[add] move Event related url processing into new controllers/ABEventController
